Change setup.py to prevent premature imports of dependencies
Previous version of setup.py assumed that numpy and Cython were already
installed when setup.py was first called. If they were not, the install
would fail even if these packages were listed as dependencies. This
approach incapsulates the imports so that they are not run before the
requirements can be met.